Move shared GUI code into AppScaffolding
