+++ Source/AppScaffolding/ObjectComparer[T].cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections;
 
-namespace LibationWinForms
+namespace AppScaffolding
 {
-	internal class ObjectComparer<T> : IComparer where T : IComparable
+	public class ObjectComparer<T> : IComparer where T : IComparable
 	{
 		public int Compare(object x, object y) => ((T)x).CompareTo(y);
 	}
