--- Source/LibationAvalonia/ObjectComparer[T].cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections;
-
-namespace LibationAvalonia
-{
-	internal class ObjectComparer<T> : IComparer where T : IComparable
-	{
-		public int Compare(object x, object y) => ((T)x).CompareTo(y);
-	}
-}
