+++ Source/LibationAvalonia/ViewModels/GridEntry.cs
@@ -1,4 +1,5 @@
 ﻿using ApplicationServices;
+using AppScaffolding;
 using Avalonia.Media;
 using DataLayer;
 using Dinah.Core;
