+++ Source/LibationAvalonia/ViewModels/ProcessQueueViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using ApplicationServices;
+using AppScaffolding;
 using Avalonia.Controls;
 using Avalonia.Threading;
 using DataLayer;
