+++ Source/LibationWinForms/ProcessQueue/ProcessQueueControl.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using ApplicationServices;
+using AppScaffolding;
 using LibationFileManager;
 
 namespace LibationWinForms.ProcessQueue
